+++ packages/kit/src/cli.js
@@ -1,8 +1,10 @@
-import sade from 'sade';
-import colors from 'kleur';
+import chokidar from 'chokidar';
 import fs from 'fs';
+import colors from 'kleur';
 import { relative } from 'path';
 import * as ports from 'port-authority';
+import sade from 'sade';
+import * as vite from 'vite';
 import { load_config } from './core/config/index.js';
 import { networkInterfaces, release } from 'os';
 import { coalesce_to_error } from './utils/error.js';
@@ -53,33 +55,96 @@ prog
 	.option('--https', 'Use self-signed HTTPS certificate')
 	.option('-H', 'no longer supported, use --https instead') // TODO remove for 1.0
 	.action(async ({ port, host, https, open, H }) => {
-		try {
-			if (H) throw new Error('-H is no longer supported — use --https instead');
+		let first = true;
+		let relaunching = false;
+		let uid = 1;
 
-			process.env.NODE_ENV = process.env.NODE_ENV || 'development';
-			const config = await load_config();
+		/** @type {() => Promise<void>} */
+		let close;
 
-			const { dev } = await import('./core/dev/index.js');
+		async function start() {
+			const svelte_config = await load_config();
+			const { plugins } = await import('./core/dev/plugin.js');
+			const vite_config = await svelte_config.kit.vite();
 
-			const cwd = process.cwd();
+			/** @type {import('vite').UserConfig} */
+			const config = {
+				plugins: [...(vite_config.plugins || []), plugins(svelte_config)]
+			};
+			config.server = {};
 
-			const { address_info, server_config } = await dev({
-				cwd,
-				port,
-				host,
-				https,
-				config
-			});
+			// optional config from command-line flags
+			// these should take precedence, but not print conflict warnings
+			if (host) {
+				config.server.host = host;
+			}
+
+			// if https is already enabled then do nothing. it could be an object and we
+			// don't want to overwrite with a boolean
+			if (https && !vite_config?.server?.https) {
+				config.server.https = https;
+			}
+
+			if (port) {
+				config.server.port = port;
+			}
+
+			const server = await vite.createServer(config);
+			await server.listen(port);
+
+			const address_info = /** @type {import('net').AddressInfo} */ (
+				/** @type {import('http').Server} */ (server.httpServer).address()
+			);
+
+			const resolved_config = server.config;
 
 			welcome({
 				port: address_info.port,
 				host: address_info.address,
-				https: !!(https || server_config.https),
-				open: open || !!server_config.open,
-				base: config.kit.paths.base,
-				loose: server_config.fs.strict === false,
-				allow: server_config.fs.allow,
-				cwd
+				https: !!(https || resolved_config.server.https),
+				open: first && (open || !!resolved_config.server.open),
+				base: svelte_config.kit.paths.base,
+				loose: resolved_config.server.fs.strict === false,
+				allow: resolved_config.server.fs.allow
+			});
+
+			first = false;
+
+			return server.close;
+		}
+
+		// TODO: we should probably replace this with something like vite-plugin-restart
+		async function relaunch() {
+			const id = uid;
+			relaunching = true;
+
+			try {
+				await close();
+				close = await start();
+
+				if (id !== uid) relaunch();
+			} catch (e) {
+				const error = /** @type {Error} */ (e);
+
+				console.error(colors.bold().red(`> ${error.message}`));
+				if (error.stack) {
+					console.error(colors.gray(error.stack.split('\n').slice(1).join('\n')));
+				}
+			}
+
+			relaunching = false;
+		}
+
+		try {
+			if (H) throw new Error('-H is no longer supported — use --https instead');
+
+			process.env.NODE_ENV = process.env.NODE_ENV || 'development';
+
+			close = await start();
+
+			chokidar.watch('svelte.config.js', { ignoreInitial: true }).on('change', () => {
+				if (relaunching) uid += 1;
+				else relaunch();
 			});
 		} catch (error) {
 			handle_error(error);
@@ -138,11 +203,10 @@ prog
 			await check_port(port);
 
 			process.env.NODE_ENV = process.env.NODE_ENV || 'production';
-			const config = await load_config();
 
 			const { preview } = await import('./core/preview/index.js');
 
-			await preview({ port, host, config, https });
+			const { config } = await preview({ port, host, https });
 
 			welcome({ port, host, https, open, base: config.kit.paths.base });
 		} catch (error) {