+++ packages/kit/src/core/config/index.js
@@ -5,6 +5,8 @@ import { logger } from '../utils.js';
 import options from './options.js';
 
 /**
+ * Loads the template (src/app.html by default) and validates that it has the
+ * required content.
  * @param {string} cwd
  * @param {import('types').ValidatedConfig} config
  */
@@ -14,7 +16,16 @@ export function load_template(cwd, config) {
 
 	if (fs.existsSync(template)) {
 		const contents = fs.readFileSync(template, 'utf8');
-		const expected_tags = ['%svelte.head%', '%svelte.body%'];
+
+		// TODO remove this for 1.0
+		const match = /%svelte\.([a-z]+)%/.exec(contents);
+		if (match) {
+			throw new Error(
+				`%svelte.${match[1]}% in ${relative} should be replaced with %sveltekit.${match[1]}%`
+			);
+		}
+
+		const expected_tags = ['%sveltekit.head%', '%sveltekit.body%'];
 		expected_tags.forEach((tag) => {
 			if (contents.indexOf(tag) === -1) {
 				throw new Error(`${relative} is missing ${tag}`);
@@ -27,6 +38,11 @@ export function load_template(cwd, config) {
 	return fs.readFileSync(template, 'utf-8');
 }
 
+/**
+ * Loads and validates svelte.config.js
+ * @param {{ cwd?: string }} options
+ * @returns {Promise<import('types').ValidatedConfig>}
+ */
 export async function load_config({ cwd = process.cwd() } = {}) {
 	const config_file = path.join(cwd, 'svelte.config.js');
 
@@ -36,7 +52,7 @@ export async function load_config({ cwd = process.cwd() } = {}) {
 		);
 	}
 
-	const config = await import(url.pathToFileURL(config_file).href);
+	const config = await import(`${url.pathToFileURL(config_file).href}?ts=${Date.now()}`);
 
 	const validated = validate_config(config.default);
 
@@ -65,6 +81,7 @@ export function validate_config(config) {
 }
 
 /**
+ * Ensures the user does not override any config values that SvelteKit must control.
  * @param {string[]} conflicts - array of conflicts in dotted notation
  * @param {string=} pathPrefix - prepended in front of the path
  * @param {string=} scope - used to prefix the whole error message