+++ packages/kit/types/ambient.d.ts
@@ -15,7 +15,15 @@
  * }
  * ```
  *
- * これらの interface を活用することで、`event.locals`、`event.platform`、`session`、`stuff` を使用する際に型の安全性を得ることができます:
+ * これらのインターフェースを作成することによって、`event.locals`、`event.platform`、`session`、`stuff` を使用する際に型の安全性を確保することができます。
+ *
+ * アンビエント宣言(ambient declaration)ファイルであるため、`import` 文を使用することができません。代わりに、`import(...)` 関数をお使いください:
+ *
+ * ```ts
+ * interface Locals {
+ * 	user: import('$lib/types').User;
+ * }
+ * ```
  */
 declare namespace App {
 	/**
@@ -41,14 +49,10 @@ declare namespace App {
 
 /**
  * ```ts
- * import { amp, browser, dev, mode, prerendering } from '$app/env';
+ * import { browser, dev, mode, prerendering } from '$app/env';
  * ```
  */
 declare module '$app/env' {
-	/**
-	 * アプリが [AMP モード](/docs/seo#manual-setup-amp) で動作しているかどうかを示します。
-	 */
-	export const amp: boolean;
 	/**
 	 * アプリがブラウザで動作しているか、それともサーバーで動作しているかを示します。
 	 */
@@ -84,7 +88,7 @@ declare module '$app/env' {
  */
 declare module '$app/navigation' {
 	/**
-	 * ナビゲーション後のページ更新の時にこれが(例えば `onMount` の中や action で)呼び出された場合、SvelteKit の組み込みのスクロール処理を無効にします。
+	 * ナビゲーション後のページ更新の時にこれが(例えば `onMount`、`afterNavigate` の中や action で)呼び出された場合、SvelteKit の組み込みのスクロール処理を無効にします。
 	 * ユーザーの期待する動きではなくなるため、一般的には推奨されません。
 	 */
 	export function disableScrollHandling(): void;
@@ -102,7 +106,7 @@ declare module '$app/navigation' {
 		opts?: { replaceState?: boolean; noscroll?: boolean; keepfocus?: boolean; state?: any }
 	): Promise<void>;
 	/**
-	 * 現在アクティブなページに属している `load` 関数が当該リソースを `fetch` する場合、再実行させます。それに続いてページが更新されたときに解決される `Promise` を返します。
+	 * 現在アクティブなページに属している `load` 関数が当該リソースを `fetch` する場合や、invalidate されたリソースがページそのものだったときにページエンドポイントからデータを再フェッチする場合に再実行させます。それに続いてページが更新されたときに解決される `Promise` を返します。
 	 * @param dependency The invalidated resource
 	 */
 	export function invalidate(dependency: string | ((href: string) => boolean)): Promise<void>;
@@ -151,7 +155,7 @@ declare module '$app/navigation' {
  */
 declare module '$app/paths' {
 	/**
-	 * [`config.kit.paths.base`](/docs/configuration#paths) にマッチする文字列です。`/` で始まる必要があります。末尾を `/` にしてはいけません。
+	 * [`config.kit.paths.base`](/docs/configuration#paths) にマッチする文字列です。先頭は `/` で始まる必要があり、末尾は `/` にしてはいけません(例 `/base-path`)。空文字(empty string)の場合はこのルールに該当しません。
 	 */
 	export const base: `/${string}`;
 	/**
@@ -279,11 +283,16 @@ declare module '@sveltejs/kit/hooks' {
 /**
  * `fetch` やそれに関連する interface の polyfill で、ネイティブ実装が提供されない環境向けの adapter が使用します。
  */
-declare module '@sveltejs/kit/install-fetch' {
+declare module '@sveltejs/kit/node/polyfills' {
 	/**
-	 * `node-fetch` を使用して、`fetch` `Headers` `Request` `Response` を global で利用できるようにします。
+	 * 様々な web API をグローバルで使用できるようにします:
+	 * - `crypto`
+	 * - `fetch`
+	 * - `Headers`
+	 * - `Request`
+	 * - `Response`
 	 */
-	export function installFetch(): void;
+	export function installPolyfills(): void;
 }
 
 /**