+++ packages/adapter-auto/CHANGELOG.md
@@ -1,5 +1,94 @@
 # @sveltejs/adapter-auto
 
+## 1.0.0-next.50
+
+### Patch Changes
+
+- Update dependencies ([#5121](https://github.com/sveltejs/kit/pull/5121))
+
+- Updated dependencies [[`821592f3`](https://github.com/sveltejs/kit/commit/821592f36e4e507cbfa5a0e0a1ee99be9991e827)]:
+  - @sveltejs/adapter-cloudflare@1.0.0-next.23
+  - @sveltejs/adapter-netlify@1.0.0-next.64
+  - @sveltejs/adapter-vercel@1.0.0-next.58
+
+## 1.0.0-next.49
+
+### Patch Changes
+
+- Update adapter entrypoint typings to be NodeNext/ESNext-compatible ([#5111](https://github.com/sveltejs/kit/pull/5111))
+
+- Updated dependencies [[`7d5fb308`](https://github.com/sveltejs/kit/commit/7d5fb3083a428c4b045d7845138efdb40e1ebad8)]:
+  - @sveltejs/adapter-cloudflare@1.0.0-next.22
+  - @sveltejs/adapter-netlify@1.0.0-next.63
+  - @sveltejs/adapter-vercel@1.0.0-next.57
+
+## 1.0.0-next.48
+
+### Patch Changes
+
+- Updated dependencies [[`60370cc2`](https://github.com/sveltejs/kit/commit/60370cc242ff0879401d3b10eb38e9c22e41b7ab), [`f0fb175d`](https://github.com/sveltejs/kit/commit/f0fb175dda834807dc51964105e4ecce33c4ad0b)]:
+  - @sveltejs/adapter-vercel@1.0.0-next.56
+  - @sveltejs/adapter-cloudflare@1.0.0-next.21
+  - @sveltejs/adapter-netlify@1.0.0-next.62
+
+## 1.0.0-next.47
+
+### Patch Changes
+
+- Add types to pkg.exports ([#5045](https://github.com/sveltejs/kit/pull/5045))
+
+- Updated dependencies [[`9b13c61c`](https://github.com/sveltejs/kit/commit/9b13c61c68ac4fe601e0a95cf958f7b63fa7a754)]:
+  - @sveltejs/adapter-cloudflare@1.0.0-next.20
+  - @sveltejs/adapter-netlify@1.0.0-next.61
+  - @sveltejs/adapter-vercel@1.0.0-next.55
+
+## 1.0.0-next.46
+
+### Patch Changes
+
+- Updated dependencies [[`e3c9073a`](https://github.com/sveltejs/kit/commit/e3c9073a8b43f1dd4aa871639b61cb974a1f110e)]:
+  - @sveltejs/adapter-netlify@1.0.0-next.60
+  - @sveltejs/adapter-vercel@1.0.0-next.54
+
+## 1.0.0-next.45
+
+### Patch Changes
+
+- Updated dependencies [[`c660143a`](https://github.com/sveltejs/kit/commit/c660143ad2b90ea3a1ae18247c553da2538750a0)]:
+  - @sveltejs/adapter-netlify@1.0.0-next.59
+  - @sveltejs/adapter-vercel@1.0.0-next.54
+
+## 1.0.0-next.44
+
+### Patch Changes
+
+- Updated dependencies [[`a452bd83`](https://github.com/sveltejs/kit/commit/a452bd83ac5feb27040a1cb364395c9ae5c04107)]:
+  - @sveltejs/adapter-vercel@1.0.0-next.53
+
+## 1.0.0-next.43
+
+### Patch Changes
+
+- Updated dependencies [[`0efa319d`](https://github.com/sveltejs/kit/commit/0efa319d0a03c180a66845d1a4933589084ddb48), [`3d34924c`](https://github.com/sveltejs/kit/commit/3d34924c3f3c1af75c326c6d92cdbfb263f0492a), [`e685c5ae`](https://github.com/sveltejs/kit/commit/e685c5aecdefbc57fcbec5624ed43fd7dd01a037)]:
+  - @sveltejs/adapter-netlify@1.0.0-next.58
+  - @sveltejs/adapter-vercel@1.0.0-next.52
+
+## 1.0.0-next.42
+
+### Patch Changes
+
+- Updated dependencies [[`7609ad07`](https://github.com/sveltejs/kit/commit/7609ad0759c895600198eaf4345d735c7c6d1166), [`7609ad07`](https://github.com/sveltejs/kit/commit/7609ad0759c895600198eaf4345d735c7c6d1166)]:
+  - @sveltejs/adapter-vercel@1.0.0-next.51
+  - @sveltejs/adapter-netlify@1.0.0-next.57
+
+## 1.0.0-next.41
+
+### Patch Changes
+
+- Updated dependencies [[`3b83249b`](https://github.com/sveltejs/kit/commit/3b83249ba22353ccfe592d265d2c3b8b71c2c224)]:
+  - @sveltejs/adapter-netlify@1.0.0-next.57
+  - @sveltejs/adapter-vercel@1.0.0-next.50
+
 ## 1.0.0-next.40
 
 ### Patch Changes